* Makefile.am (libgmp.la, libmp.la): Revert change to build from 	mpn/libmpn.la etc, go back to explicitly listed objects.
Well this was an exercise in futility.  The comments that were there
before about the use of mpn/libmpn.la etc putting PIC objects into
non-PIC libraries remained true, if I'd bothered to check.

The use of mpn/libmpn.la etc also seems to cause problems on HP-UX 9
and 10, possibly related to parsing the output of "ar t" on those
systems, but in any case causing a couple of mpn objects to be omitted
from the final libgmp (.so and .a).