* gmp-h.in (__GMPN_ADD_1, __GMPN_SUB_1): New macros, rewrite of 	mpn_add_1 and mpn_sub_1, better code for src==dst and/or n==1, 	separate versions for gcc x86, gcc generic, and non-gcc. 	(mpn_add_1, mpn_sub_1): Use them. 	(mpn_add, mpn_sub): Ditto, to get inlines on all compilers. 	(extern "C") [__cplusplus]: Let this encompass the extern inlines too.
	* gmp-h.in (__GMP_LSYM_PREFIX): New substitution.
	(__GMP_ASM_L): New macro.

The main gain in this new mpn_add_1 and mpn_sub_1 is optimizing
compile-time src==dst to get rid of the copy loops.  mul_fft.c is a
big (ab)user of mpn_add_1 and mpn_sub_1 and the code size there on k6
shrinks by no less than 1400 bytes.

The x86 nonsense is perhaps a touch excessive for the few instructions
it saves, but if we know what we want to see on those chips and know
how to get it then why hold back.